requirements.txt
django-1.6.4
rdflib-4.1.2
rdflib-sqlite
pyparsing<=1.5.7
Pyrex
django-authority
itsdangerous
pip install git+https://github.com/RDFLib/rdflib-leveldb.git@master
